# public/openapi.yaml
openapi: 3.0.3
info:
  title: E-Puskesmas EIS API
  version: 1.0.0
  description: |
    OpenAPI untuk E-Puskesmas EIS. Disusun mengikuti struktur menu frontend:
    Dashboard, Cari, Klaster 1-4, Lintas Klaster, Monitoring, dan Akun.

    **Integrasi pihak ketiga (e-Puskesmas / webhook)**
    - Semua pengiriman data dari sistem eksternal dilakukan via endpoint `POST /api/eis/klaster1`, `POST /api/eis/klaster2`, `POST /api/eis/klaster3`, `POST /api/eis/klaster4`, dan `POST /api/eis/lintas-klaster`.
    - Endpoint dashboard/analitik tetap `GET` (read-only). Data yang tampil berasal dari hasil POST (webhook push).
    - Semua endpoint POST di atas wajib menyertakan header `X-Epus-Signature` untuk HMAC:
      signature = HMAC_SHA256(EPUS_WEBHOOK_SECRET, `${t}.${rawJsonUtf8}`)
      Header format: `t=<unix_ts>,v1=<hex_sha256>`
    - Idempotency: gunakan `batchId` unik agar retry aman (tidak dobel).
    - Skala besar: payload boleh `Content-Encoding: gzip` (signature dihitung dari JSON *setelah* di-decompress).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    XEpusSignature:
      name: X-Epus-Signature
      in: header
      required: true
      description: "Signature HMAC. Format: `t=<unix_ts>,v1=<hex_sha256>`"
      schema:
        type: string
        example: t=1700000000,v1=0123abcd...
    ContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: "Opsional untuk payload besar. Gunakan `gzip` bila body di-compress."
      schema:
        type: string
        enum: [gzip]
...
    PuskesmasId:
      name: puskesmasId
      in: query
      required: false
      schema:
        type: string
      description: "ID puskesmas, gunakan `all` untuk semua"
    Bulan:
      name: bulan
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
      description: Bulan (1-12)
    Tahun:
      name: tahun
      in: query
      required: false
      schema:
        type: integer
      description: Tahun (YYYY)
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    KlasterGenericResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          additionalProperties: true
...
paths:
...
  /api/eis/klaster4:
    get:
      tags: [Klaster 4]
      summary: Data Klaster 4
      parameters:
        - $ref: '#/components/parameters/PuskesmasId'
        - $ref: '#/components/parameters/Bulan'
        - $ref: '#/components/parameters/Tahun'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Data klaster 4
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlasterGenericResponse'
    post:
      tags: [Klaster 4]
      summary: Webhook push data Klaster 4 (Top Diagnosa, Diagnosa Harian)
      description: |
        Dipakai untuk mengirim data Klaster 4.

        Minimal kirim salah satu:
        - `topDiagnosa` (bulanan, untuk grafik Top Diagnosa)
        - `diagnosaHarian` (harian, untuk klasifikasi bahaya tinggi/sedang/rendah)

        Untuk `diagnosaHarian`, server akan mengganti (replace) semua record pada tanggal-tanggal yang dikirim.
      parameters:
        - $ref: '#/components/parameters/XEpusSignature'
        - $ref: '#/components/parameters/ContentEncoding'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId: { type: string }
                puskesmasKode: { type: string }
                puskesmasId: { type: string }
                bulan: { type: integer, minimum: 1, maximum: 12 }
                tahun: { type: integer }
                topDiagnosa:
                  type: array
                  items:
                    type: object
                    required: [icd10Code, diagnosa, jumlahKasus]
                    properties:
                      icd10Code: { type: string }
                      diagnosa: { type: string }
                      jumlahKasus: { type: integer }
                diagnosaHarian:
                  type: array
                  items:
                    type: object
                    required: [tanggal, icd10Code, diagnosa, jumlahKasus, tingkatBahaya]
                    properties:
                      tanggal: { type: string, format: date }
                      icd10Code: { type: string }
                      diagnosa: { type: string }
                      jumlahKasus: { type: integer }
                      tingkatBahaya:
                        type: string
                        enum: [TINGGI, SEDANG, RENDAH]
              oneOf:
                - required: [batchId, puskesmasKode, bulan, tahun]
                - required: [batchId, puskesmasId, bulan, tahun]
            examples:
              only_top_diagnosa:
                summary: Kirim Top Diagnosa saja (bulanan)
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K4-TOP-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  topDiagnosa:
                    - icd10Code: J06.9
                      diagnosa: Infeksi saluran pernapasan atas akut, tidak spesifik
                      jumlahKasus: 120
                    - icd10Code: A09
                      diagnosa: Diare dan gastroenteritis yang diduga berasal dari infeksi
                      jumlahKasus: 65
              only_diagnosa_harian:
                summary: Kirim Diagnosa Harian saja (harian; replace per tanggal yang dikirim)
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K4-HARIAN-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  diagnosaHarian:
                    - tanggal: 2026-02-01
                      icd10Code: A09
                      diagnosa: Diare
                      jumlahKasus: 10
                      tingkatBahaya: SEDANG
                    - tanggal: 2026-02-01
                      icd10Code: J06.9
                      diagnosa: ISPA
                      jumlahKasus: 22
                      tingkatBahaya: RENDAH
                    - tanggal: 2026-02-02
                      icd10Code: I10
                      diagnosa: Hipertensi esensial (primer)
                      jumlahKasus: 8
                      tingkatBahaya: TINGGI
      responses:
        '200':
          description: Data diterima dan diproses
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Signature invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Kesalahan server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...